SOL-151286: make panic recovery unconditional (drop OBS_PANIC_RECOVERY_ENABLED), move to internal/middleware/recovery (amitmorade review)
