--- internal/observability/recovery/recovery_test.go
@@ -1,33 +1,0 @@
-// Copyright 2024-2026 Solace Corporation. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package recovery
-
-import (
-	"testing"
-
-	"github.com/SolaceDev/solace-broker-mcp/internal/config"
-)
-
-// TestEnabled pins that Enabled reflects the PanicRecoveryEnabled flag rather
-// than a hardcoded constant — both directions.
-func TestEnabled(t *testing.T) {
-	t.Parallel()
-	for _, want := range []bool{true, false} {
-		cfg := config.ObservabilityConfig{PanicRecoveryEnabled: want}
-		if got := Enabled(cfg); got != want {
-			t.Errorf("Enabled() = %v, want %v", got, want)
-		}
-	}
-}
